message editing issues bypass
