+++ anketa.js
@@ -4,41 +4,14 @@ import { searchForNew } from "./filedata.js";
 import { dataBot } from './values.js';
 import { writeGoogle, readGoogle } from './crud.js';
 import { checkStatus } from './interval.js';
+import { phrases, keyboards } from './language_ua.js'
 
 let customerPhone;
 let customerName;
 let customerInfo = {};
 let selectedOrderRaw;
-
-
 const spreadsheetId = dataBot.googleSheetId;
 const phoneRegex = /^\d{10,12}$/;
-
-const phrases = {
-  greetings: 'Привіт, якщо ви хочете зробити замовлення, натисніть кнопку "Зробити замовлення".',
-  contactRequest: 'Нам потрібні ваші контактні дані. Отримати з контактних даних телеграм?',
-  dataConfirmation: `Ваш номер телефону: ${customerPhone}. Ваше імя ${customerName}. Дані вірні?`,
-  thanksForOrder: `Замовлення успішно оформлено. Дякую ${customerName}`,
-  wrongName: 'Невірне ім\'я. Будь ласка, введіть своє справжнє ім\'я:',
-  wrongPhone: 'Невірний номер телефону. Будь ласка, введіть номер телефону ще раз:',
-  phoneRules: 'Введіть ваш номер телефону без +. Лише цифри. І відправте повідомлення',
-  nameRequest: 'Введіть своє ім\'я:',
-};
-
-const keyboards = {
-  startingKeyboard: [['Зробити замовлення']],
-  contactRequest: [
-    [ { text: 'Так', request_contact: true, } ],
-    ['Ні, я введу номер вручну'],
-    ['/start'],
-  ],
-  dataConfirmation: [
-    ['Так, Оформити замовлення'],
-    ['Ні, повторити введення'],
-    ['/start'],
-  ],
-  enterPhone: [ ['/start'] ]
-}
 
 export const anketaListiner = async() => {
     bot.setMyCommands([
@@ -72,7 +45,6 @@ export const anketaListiner = async() => {
     });
     
     bot.on('message', async (msg) => {
-      console.log(customerInfo);
       const chatId = msg.chat.id;
       if (msg.text === 'Зробити замовлення') await sendNewRowsToTelegram(spreadsheetId, dataBot.googleSheetName, dataBot.statusColumn, chatId, bot);
       else if (msg.contact) {
@@ -83,7 +55,7 @@ export const anketaListiner = async() => {
         customerInfo[chatId].phone = msg.contact.phone_number;
         customerPhone = msg.contact.phone_number;
         customerName = msg.contact.first_name;
-        bot.sendMessage(chatId, `Ваш номер телефону: ${customerInfo[chatId].phone}. Ваше імя ${customerInfo[chatId].name}. Дані вірні?`, { 
+        bot.sendMessage(chatId, phrases.dataConfirmation(customerInfo[chatId].phone, customerInfo[chatId].name), { 
           reply_markup: { keyboard: keyboards.dataConfirmation, resize_keyboard: true, one_time_keyboard: true }});
       } else if(msg.text === 'Так, Оформити замовлення') {
           const chatId = msg.chat.id;
@@ -92,16 +64,21 @@ export const anketaListiner = async() => {
             await writeGoogle(`${dataBot.googleSheetName}!${dataBot.statusColumn}${customerInfo[chatId].lotNumber}`, [['done']]);
             await writeGoogle(`${dataBot.googleSheetName}!${dataBot.user.nameColumn}${customerInfo[chatId].lotNumber}`, [[customerInfo[chatId].name]]);
             await writeGoogle(`${dataBot.googleSheetName}!${dataBot.user.phoneColumn}${customerInfo[chatId].lotNumber}`, [[customerInfo[chatId].phone]]);
-            const message_id = await (await readGoogle(`${dataBot.googleSheetName}!${dataBot.content.message_idColumn}${customerInfo[chatId].lotNumber}`))[0];
-            const oldMessage = await readGoogle(`${dataBot.googleSheetName}!${dataBot.content.startColumn}${customerInfo[chatId].lotNumber}:${dataBot.content.endColumn}${selectedOrderRaw}`);
-            const oldMessageString = oldMessage.join('\n');
-            const newMessage = "📌 " + oldMessageString;
-            try {
-              await bot.editMessageText(newMessage, {chat_id: dataBot.channelId, message_id: message_id});
-            } catch (error) {
-              console.log(error)
-            }
-            bot.sendMessage(chatId, `Замовлення успішно оформлено. Дякую ${customerInfo[chatId].name}`);
+            const editingMessage = async () => {
+              const message_id = await (await readGoogle(`${dataBot.googleSheetName}!${dataBot.content.message_idColumn}${customerInfo[chatId].lotNumber}`))[0];
+              const oldMessage = await readGoogle(`${dataBot.googleSheetName}!${dataBot.content.startColumn}${customerInfo[chatId].lotNumber}:${dataBot.content.endColumn}${selectedOrderRaw}`);
+              const oldMessageString = oldMessage.join('\n');
+              const newMessage = "📌 " + oldMessageString;
+              if (message_id) {
+                try {
+                  await bot.editMessageText(newMessage, {chat_id: dataBot.channelId, message_id: message_id});
+                } catch (error) {
+                  
+                }
+              } 
+            };
+            editingMessage();
+            bot.sendMessage(chatId, phrases.thanksForOrder(customerInfo[chatId].name));
           } 
       } else if (msg.text === 'Почати спочатку') {
         bot.sendMessage(chatId, '/start');
@@ -119,7 +96,7 @@ export const anketaListiner = async() => {
         if (msg.text.length >= 2) {
         customerName = msg.text;
         customerInfo[chatId].name = msg.text;
-        bot.sendMessage(chatId, `Ваш номер телефону: ${customerInfo[chatId].phone}. Ваше імя ${customerInfo[chatId].name}. Дані вірні?` , {
+        bot.sendMessage(chatId, phrases.dataConfirmation(customerInfo[chatId].phone, customerInfo[chatId].name), {
           reply_markup: { keyboard: keyboards.dataConfirmation, resize_keyboard: true, one_time_keyboard: true },
         });
         };
